backend2/gRPC/requirements.txt
grpcio
grpcio grpcio-tools

flask
redis
pyjwt
fastapi==0.68.0
uvicorn==0.15.0
passlib==1.7.4
sqlalchemy==1.4.0
python-multipart
pymongo
aiofiles